compressionparams: force signed compare to avoid compiler warning
GCC detects the <0 comparison with an unsigned and emits a warning.
So cast to a signed integer to silence it.

With this change, our C code is -Wall -Wextra -Wpedantic clean
except for -Wmissing-field-initializers and -Wunused-parameters.
We could fix -Wmissing-field-initializers, but it is a lot of extra
typing to fill in the Python data structures. Most of the
-Wunused-parameters warnings are with regards to extra
PyObject* arguments. We can't just drop those. So I'm fine
with ignoring that warning.

+++ c-ext/compressionparams.c
@@ -106,7 +106,8 @@ static int CompressionParameters_init(CompressionParametersObject* self, PyObjec
 		return -1;
 	}
 
-	if (strategy < ZSTD_fast || strategy > ZSTD_btopt) {
+    /* ZSTD_fast is 0 and compiler may complain about `< 0` check with unsigned. */
+	if ((int)strategy < ZSTD_fast || strategy > ZSTD_btopt) {
 		PyErr_SetString(PyExc_ValueError, "invalid strategy value");
 		return -1;
 	}
